+++ lemminx-maven/src/main/java/org/eclipse/lemminx/extensions/maven/participants/completion/MavenCompletionParticipant.java
@@ -97,12 +97,10 @@ import org.eclipse.lemminx.dom.DOMDocument;
 import org.eclipse.lemminx.dom.DOMElement;
 import org.eclipse.lemminx.dom.DOMNode;
 import org.eclipse.lemminx.extensions.maven.DOMConstants;
-import org.eclipse.lemminx.extensions.maven.DependencyScope;
 import org.eclipse.lemminx.extensions.maven.MavenInitializationException;
 import org.eclipse.lemminx.extensions.maven.MavenLemminxExtension;
 import org.eclipse.lemminx.extensions.maven.MavenModelOutOfDatedException;
 import org.eclipse.lemminx.extensions.maven.MojoParameter;
-import org.eclipse.lemminx.extensions.maven.Phase;
 import org.eclipse.lemminx.extensions.maven.participants.ArtifactWithDescription;
 import org.eclipse.lemminx.extensions.maven.searcher.RemoteCentralRepositorySearcher.OngoingOperationException;
 import org.eclipse.lemminx.extensions.maven.utils.DOMUtils;
