+++ lemminx-maven/src/main/java/org/eclipse/lemminx/extensions/maven/settings/XMLMavenCentralSettings.java
@@ -9,7 +9,7 @@
  * Contributors:
  * Red Hat Inc. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.lemminx.extensions.maven;
+package org.eclipse.lemminx.extensions.maven.settings;
 
 import java.util.Objects;
